requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
python-docx
google-generativeai
python-multipart